+++ Staley RC 2019/src/main/java/frc/robot/Robot.java
@@ -12,7 +12,7 @@ import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.Scheduler;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.commands.ControllerDrive;
+import frc.robot.commands.drivetrain.ControllerDrive;
 
 /**
  * The VM is configured to automatically run this class, and to call the
